tools/skb-cli/pyproject.toml
[project]
name = "skb-cli"
version = "1.0.0"
description = "Shared Knowledge Base CLI - Enterprise Knowledge Graph for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Shared KB Curator", email = "curator@team.com"}
]
keywords = ["claude-code", "knowledge-base", "cli", "enterprise"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "gitpython>=3.1.40",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "semver>=3.0.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[project.scripts]
sku = "sku.cli:main"

[project.urls]
Homepage = "https://github.com/your-team/shared-knowledge-base"
Documentation = "https://github.com/your-team/shared-knowledge-base/blob/main/tools/skb-cli/README.md"
Repository = "https://github.com/your-team/shared-knowledge-base"
"Bug Tracker" = "https://github.com/your-team/shared-knowledge-base/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sku"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
